BeanFactory 리펙토링
