+++ src/main/java/core/di/factory/BeanFactory.java
@@ -3,19 +3,24 @@ package core.di.factory;
 import com.google.common.collect.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.BeanUtils;
 
+import java.lang.reflect.Constructor;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public class BeanFactory {
     private static final Logger logger = LoggerFactory.getLogger(BeanFactory.class);
 
-    private final PreInstantiateBeans preInstantiateBeans;
-
     private Map<Class<?>, Object> beans = Maps.newHashMap();
 
+    private Set<Class<?>> preInstantiateBeans;
+
     public BeanFactory(Set<Class<?>> preInstantiateBeans) {
-        this.preInstantiateBeans = new PreInstantiateBeans(preInstantiateBeans);
+        this.preInstantiateBeans = preInstantiateBeans;
     }
 
     @SuppressWarnings("unchecked")
@@ -24,6 +29,36 @@ public class BeanFactory {
     }
 
     public void initialize() {
-        beans.putAll(preInstantiateBeans.createBeansObject());
+        beans.putAll(preInstantiateBeans());
+    }
+
+    public Map<Class<?>, Object> preInstantiateBeans() {
+        return preInstantiateBeans.stream()
+                .collect(Collectors.toMap(b -> b, this::instantiateClass));
+    }
+
+    public Object instantiateClass(Class<?> clazz) {
+        try {
+            if (beans.containsKey(clazz)) {
+                return beans.get(clazz);
+            }
+            Class<?> conClass = BeanFactoryUtils.findConcreteClass(clazz, preInstantiateBeans);
+            Constructor<?> constructor = BeanFactoryUtils.getInjectedConstructor(clazz);
+            if (constructor == null) return conClass.newInstance();
+            logger.debug("{}", constructor);
+            return instantiateConstructor(constructor);
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    private Object instantiateConstructor(Constructor<?> constructor) {
+        Class<?>[] parameterTypes = constructor.getParameterTypes();
+        List<Object> args = new ArrayList<>();
+        for (Class<?> clazz : parameterTypes) {
+            Object bean = instantiateClass(clazz);
+            args.add(bean);
+        }
+        return BeanUtils.instantiateClass(constructor, args.toArray());
     }
 }
